test(topic-fence): consolidate tests into session-extract.test.ts
Upstream CONTRIBUTING.md prohibits creating new test files without
prior maintainer approval. Merge the topic-fence test files into
the existing session-extract.test.ts under clearly-labeled
"topic-fence:" describe blocks.

- Remove tests/session/topic-fence.test.ts
- Remove tests/session/topic-fence-drift.test.ts
- Append all 50 tests to tests/session/session-extract.test.ts

No behavior change; test coverage preserved.
